Move from Django to FastAPI/Flask
Step 3 - Make it launch
