--- core/MedsRecognition/supabase_client.py
@@ -1,9 +1,0 @@
-import os
-
-from supabase import create_client, Client
-
-
-def get_supabase_client() -> Client:
-    url = os.environ.get("SUPABASE_URL")
-    key = os.environ.get("SUPABASE_KEY")
-    return create_client(url, key)
