--- core/MedsRecognition/wsgi.py
@@ -1,6 +1,0 @@
-import os
-from django.core.wsgi import get_wsgi_application
-
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "MedsRecognition.settings")
-
-application = get_wsgi_application()
