+++ core/auth_service.py
