+++ core/main.py
@@ -1,37 +1,36 @@
-from sqlalchemy import create_engine
-from sqlalchemy.pool import NullPool
-from dotenv import load_dotenv
-from fastapi import FastAPI
-from fastapi.templating import Jinja2Templates
-
 import os
 
-# Load environment variables from .env
-load_dotenv()
+from dotenv import load_dotenv
+from fastapi import FastAPI
+from sqlalchemy import create_engine
+from sqlalchemy.pool import NullPool
 
-# Fetch variables
-USER = os.getenv("DATABASE_USER")
-PASSWORD = os.getenv("DATABASE_PASSWORD")
-HOST = os.getenv("DATABASE_HOST")
-PORT = os.getenv("DATABASE_PORT")
-DBNAME = os.getenv("DATABASE_NAME")
+from app_entry import AppEntry
 
-# Construct the SQLAlchemy connection string
-DATABASE_URL = f"postgresql+psycopg2://{USER}:{PASSWORD}@{HOST}:{PORT}/{DBNAME}?sslmode=require"
 
-# Create the SQLAlchemy engine
-# engine = create_engine(DATABASE_URL)
-# If using Transaction Pooler or Session Pooler, we want to ensure we disable SQLAlchemy client side pooling -
-# https://docs.sqlalchemy.org/en/20/core/pooling.html#switching-pool-implementations
-engine = create_engine(DATABASE_URL, poolclass=NullPool)
+def load_db():
+    # Load environment variables from .env
+    load_dotenv()
 
-# Test the connection
-try:
-    with engine.connect() as connection:
-        print("Connection successful!")
-except Exception as e:
-    print(f"Failed to connect: {e}")
+    # Fetch variables
+    USER = os.getenv("DATABASE_USER")
+    PASSWORD = os.getenv("DATABASE_PASSWORD")
+    HOST = os.getenv("DATABASE_HOST")
+    PORT = os.getenv("DATABASE_PORT")
+    DBNAME = os.getenv("DATABASE_NAME")
 
-templates = Jinja2Templates(directory="templates")
+    DATABASE_URL = f"postgresql+psycopg2://{USER}:{PASSWORD}@{HOST}:{PORT}/{DBNAME}?sslmode=require"
+    engine = create_engine(DATABASE_URL, poolclass=NullPool)
+
+    # Test the connection
+    try:
+        with engine.connect() as connection:
+            print("DB connection established successfully!")
+    except Exception as e:
+        print(f"Failed to connect to DB: {e}")
+    pass
+
 
 app = FastAPI()
+load_db()
+entry_point = AppEntry(app=app)
