+++ core/medication_serializer.py
