+++ core/ocr_service.py
