+++ core/requirements.txt
@@ -1,14 +1,11 @@
 alembic
-#Django==4.2.16
-#djangorestframework==3.15.2
 easyocr==1.7.2
 gevent==24.11.1
 gunicorn==22.0.0
-#Pillow==11.1.0
 psycopg2-binary==2.9.10
 pydantic
 python-dotenv
-#python-decouple==3.8
+python-multipart
 Requests==2.32.3
 sqlalchemy
 supabase==2.11.0
