+++ core/static/favicon.ico
