+++ core/templates/dashboard.html
