+++ core/templates/logout.html
