+++ core/templatetags/base64_filters.py
